refactor: break down task protocol.rs into focused modules
- Split 1721-line protocol.rs into organized module hierarchy
- Largest file now 236 lines (down from 1721)
- Group by function: client, provider, handlers, types
- Use directory hierarchies, no underscores
- Maintain all functionality and test coverage
